usb: fix ehci isochronous "in" split transactions
For "in" transactions, the "Total Bytes to Transfer" field in
the siTD is decremented by the controller by the actual transfer
size, so what remains in the field is the residue number of bytes.

Also, handle restart when we get a zero byte read on a isochronous
"in" endpoint in devusb itself (removing the restart code for
xhci drivers).

This fixes audio recording with a usb1.1 audio device connected
to ehci controller.