patch: apply patches atomically
When applying a patch, we used to write out the patch,
rename the file, and then move to the next one. This
meant that if a patch failed to apply cleanly, we would
leave the patch half-applied.

This sucked, so we now apply the patch as a whole unit,
or not at all.