rcpu: show error when we auth but don't connect
When a user isn't allowed to log into a file server, we
used to silently dump the user back to the console,
reporting that the command succeeded.

We should give the user an error when they fail to log in,
and we should exit unsuccessfully.

This status is communicated by setting up a sentinel env
var, and clearning it just before we execute the user
command.