cdfs: handle block sizes correctly
The Readblock enum does not work when you try to read audio cds.
4 cdda blocks are lager than 8192 and the command fails (at least
when using an usb cd drive).  This sets the block count according
to the track block size.  With this and the fix for libdisk I can
rip audio cds on my el cheapo usb dvd drive.  Tested with
"Midnight Oil - Blue Sky Mining"