upas/fs: clear errstr after chkunix
The readmessage loop clears errstr at start and expects it not to
change unless there is a read error.  However, strtotm may use
multiple calls to tmparse while trying to determine the date format,
which may leave errstr non-empty on success.  Clear it after chkunix
so that this doesn't get treated as a message read error.

This also fixes handling of naked From lines, which were previously
just warned about, but since the conversion to tmparse they
accidentally triggered a message read failure.