nusb/audio: enumerate streams through control interface
USB audio 2.0 splits the sample rate control from the AudioStreaming
alternate settings to a clock entity connected to input/output
terminals associated with the AudioStreaming interfaces. So, in order
to pair endpoints with their corresponding clock, we must have a
better picture of the audio function topology.

The AudioControl interface tells us which streams are available, so we
can go through each one, determining the possible stream parameters
and locating the associated endpoint. This will give us the necessary
context to locate the corresponding terminal (and soon, clock). When
we go to setup input/output endpoints, we now only have to consider
those that we have selected earlier.