imx8: cleanup framebuffer code
We can get rid of a bunch of stuff now that we'r using a softscreen.

Get rid of the static Memimage and Memdata for the framebuffer;
just allocate gscreen as a memimage from the image pool.

Avoid exporting fbraw pointer, just have screeninit() return it.

Remove fbmemalloc() function. it used to create a write-thru
mapping but given that we are no more reading the framebuffer,
we can use ucalloc() instead.