sdmmc: multiple controller support
remove the global sdio struct and add a addmmcio() function
for drivers to register their controllers.

there is a SDio.aux pointer now where controller drivers
can stash a pointer to their private data.

SDio.init() also can signal that it wants to be called
again for more controllers by returning > 0 (see pc/pmmc.c).

this is in preparation for supporting the internal mmc
device in the mnt-reform which is hooked up to usdhc1.