nusb/kb: GAOMON S620 tablet support
The GAOMON S620 tablet is recognized as a HID device,
but is restricted in the X-axis to the width of a mobile
phone.

I sniffed usb traffic for generic windows 10 hid driver
and there didnt seem anything out of the ordinary.

It turns out that this is some kind of phone feature
and i suspect they have some heuristic for how windows
reads the device and config descriptors to decide if
this is windows or android.

Checking the DIGImend project git repository, they suggest
this is some kind of UCLOGIC compatible tablet, which
once a series of string properties have been read will
switch itself to some different mode.

The report descriptor is supposed to be generated from
the string properties, but the report format didnt really
match theirs.

So i ignore the string properties and just looked at the
report data.

The format after reading the magic string properties
seems to be:

08 - report id
bb - buttons: 0x80 = inrange, 0x04 = eraser, 0x02, barrel, 0x01 = touch
xx xx - x-axis: [0x0000-0x80000)
yy yy - y-axis: [0x0000-0x50000)
pp pp - pressure: [0x0000-0xffff]
?? ?? - unknown

So i hand rolled a hid report descriptor for this and
call it a day :)