sdiahci, sdodin, sdide, devsd: get it right...
previous commit was typoed, and the changes are even wrong.
we actually need to allocate individual SDev's for the
unregistration to happen.

also, sdide's ataclear() was trying to free sdev->name,
which isnt even dynamically allocated. the clear routine
should only get rid of the controller specific stuff,
not touch the sdev at all.

devsd was also not freeing the unitflag and unit pointer
arrays...