ip/ipconfig: remove non-existant ip addresses from /net/ndb, check link-local adress in ra6 daemon
when updating /net/ndb, drop entries for
non-existant ip addreses. this prevents
accumulating alot of non-temporary dhcp
ip addresses that have been dropped when
the interface was unbound.

for the ra6 daemon, always check that the
interface still has our link-local address.

reuse Ipifc's for ipv6. we have now two
Ipifc pointers:

Ipifc *allifcs;
Ipifc *myifc;

the allifcs is updated by findmyifc() and
putndb() and contains all the interfaces
on the stack, while myifc is for our interface.
both pointers are distinct and must be
refreshed separately with readipifc().