imx8: dont reuse the TTRB1 page tables for TTRB0 identity map, remove gpioinit()
instead of putting identity map into the L1,
just allocate one more page (L1BOT) for top
level page table dedicated for the identity map
(used during mmuenable and mp startup and
reboot), then we can just load TTRB0 with it
when we need it wihout all that tlb flushing.

wastes more memory, but it is the correct thing
todo...

this also makes 4K pages work as there was a
conflict with the entries between the identity
map and the uncached memory.

also instead of unmapping stuff beyond initmap,
we can just not map anything beyond INITMAP.

that gets the kernel up fine and meminit()
will later map the rest.

i will probably backport this to the bcm64
kernel later...

for gpio, remove gpioinit and just put that
in a link function.

do the lcdinit() as soon as possible...