stats(8): drop 802.11, consider *ALL* /net/ether*/stats
The 802.11 signal strength was specific to the
pcmcia wavelan driver. I do not think anyone
has one at this point.

Ethernet used to only consider /net/ether0/stats file,
but with usb ethernets or multiple ethernet devices
it seems misleading and arbitrary.

Instead, sum up all the statistics counters of
all ethernet devices found and preset that for the graph.