nusb/ptp: transaction IDs are sequences of numbers starting with 0 (thanks unobe)
According to 'Digital Camera connectivity solutions using the picture
transfer protocol' [1] Section 2.3: "the PTP is specified using the
transaction model.  A transaction is composed of a request operation,
followed by an optional data transfer and a response.  Each
transaction has an identifier (TransactionID) that is session unique
and comprise of a 32 bit unsigned number.  The transaction IDs are a
sequence of numbers starting with 0x00000000 (for the SessionOpen
transaction) and increasing with every following operation."

[1] https://people.ece.cornell.edu/land/courses/ece4760/FinalProjects/f2012/jmv87/site/files/PTP%20Protocol.pdf