/sys/src/cmd/?i: correct memory access
Syscall handlers were not converting long's to
vlong's in a portable manner, failing on little endian.

Segments were not created w.r.t. segment allignment, which
now differs from the page size on mips and power. DATA reads
need to then also allow for accessing data between the end of
the defined DATA segment from the source text, and the end of
the segment.