kernel: zero up->ureg on sysexec() and pexit()
The up->ureg pointer points to the last delivered
note in the user-space stack.

On exec(), all the notes are cleared and the
ureg pointer should be zeroed as well as it
points into the previous programs stack.

Note, this can only happen if someone does
exec() from a note handler.

When we pexit(), also zero the up->ureg and up->dbgreg
as we are about to release the memory segments
and nobody should debugging user-space anymore.