pc64: use chain of IDIVQ for delayloop() (former aamloop())
On modern machines, doing a empty loop
is too fast, resulting in us using the
maximum loopconst of 1000000 and getting
wrong delay() timings.

To fix the timings, use a chain of IDIVQ
instructions instead.

The loop timings is going to be measured
using the TSC, so the exact timing doesnt
need to be known.

Rename aamloop() to delayloop() as it does
different things depending on 386 or amd64.