ip/torrent: fix webseed
Commit 5d1ba1a79297c47c9852e91c54d3ea8db5cff685 broke
webseeding with the last block.

The haveiece() call at the end was because the inner
is not calling havepiece() on the last block as it
does not take the piece length into account.

Now, instead, fix the inner loop, making the code
more setright foward so we call havepiece() on the
last block.