devusb: fix TT properties, checking, and bugs
When a low/full speed device is connected to a USB2.0 hub,
the USB2.0 hub needs to be send special split transaction
protocol messages to communicate with the device below.

This also applies if the fullspeed/lowspeed device is not
directly connected to the USB2.0 hub, but has a fullspeed
hub in between like:

rootport -> usb2.0 hub -> usb1.1 hub -> fs/ls device

In this case, the tranaction translator is actually
the first hub, nut the direct parent of the device.

This was all totally wrong in the hci drivers drivers.

Also, with the new interface, usbd passes number of
ports, and TT properties in the "hub" ctl message,
so the port-count, TT Think-Time and Multi-TT properties
can be properly applied by the xhci driver.

Another bug was that the xhci route string was not
correct if a hub has more than 15 ports. A USB2.0
hub can have more than 15 ports and the standard sais
that in this case a value of 15 should be used in
its bit routestring nibble.

Also, check the hub depth. We should not exceed more
than 5 hubs.

For xhci, it turns out the hub parameters where
actually never properly applied, as the spec sais
only the first "create endpoint" command applies
the hub parameters. The "evaluate context" command
does not work.

Some pikeshedding in devusb:

- fix the freaking locking.
- remove redundant parameters (isroot -> depth < 0, ishub -> nports > 0)
- add TT properties to usb device struct

With these changes, the wired "middle port" issues
on mnt-reform xhci are gone.