devusb: destruct usb tree from the leaves, split epclose() into epstop()/epclose()
For xhci, we want to keep the hubs around until all its
attached devices have been released, so have the Udev
take a reference to its parent hubs ep0.

This also means that we can now use just a pointer
to the tthub instead of duplicating the properties
needed for xhci and ythe code becomes trivial.

Do a non-recursive implementation of putep() to
conserve stack space.

For device detaches, we want to immediately cancel
all I/O so that the driver can release the device
as soon as possible.

For this, we add epstop callback in the Hci struct.