kernel: reorganize bootfs.paq generation
We do not want to unconditionally pull in all of
/lib/firmware, when we dont have the driver for it.

So group the files into separate proto-files
and make boot/bootmkfile simply checks that the basename
of a proto file matches the basename in $OBJ to be
considered.

This way, we can also avoid including user-space usb
daemon when the kernel lacks the usb driver.

Same for machines without devsd, devfs or devusb,
theres no disk so no reason to include disk file
servers.

The bootfs.paq file is specific to the configuration,
so give it a unique name. (Just using bootfs.paq
is fine as long as theres only a single configuration,
but i want to avoid future mistakes as people copy-paste
the kernel configuration).