audio/mixfs: set and forward a more sensible audio delay amount
remove the hardcoded constant which does not honor what the underlying
driver sets or uses.  instead, set a default value ourselves, the same
as what the drivers already use, and update it on change.  this fixes
issues with the emulators where the delay being too small caused a
high amount of jitter in the time it took to write the audio buffer,
severely impacting performance.