gefs: allow large snapshot names in gefs
when scanning the snapshot name table, a long snapshot would
trigger an error copying the key-value pairs, and we could
mutate the key we were looking up, leading to finding the
wrong value; all other places where the user can control the
input uses Kvmax for the size of the return buffer, so we
should use that here. We should also separate the input and
output variables in btlookup.