kernel/arm64: fix procsetup() TPIDR_EL0 potentially not getting reset to 0 on exec
We have to be carefull with the order here.

Imagine we get preepted immediately after we set p->tpidr = 0,
procsave() would override up->tpidr again with the current
register value. once we get switched back in, we then
execute syswr(TPIDR_EL0, p->tpidr), failing to reset the
register to 0.

So the correct order is to first set the register, and
then we can (optionally) set p->tpidr = 0 (doesnt matter).