cc: OLABEL and OCASE nodes do not chain on Node.right (fixes case range warning)
The code appers to have had some unused feature
where OLABEL and OCASE nodes where threated like
OLIST/OCOMMA where after generation it would chain
on the Node.right pointer to the next one.

Hoever, this appears to be unused. And when compiling
OCASE nodes with a range then Node.right points
to the end-range constant, which we should not
compile (it left a value not used error).