rc: fix fn sigexit handling (thanks nepomuk)
The sigexit pseudo note handler was only ran when we
did a explicit exit call, not when returning from
the last thread.

Also, the exitnext() check in Xsimple() that  avoids
an extra fork/wait must take into account if we need
to run the sigexit handler, so add a trapexit() function
that tells if we need to run the handler and skip
the optimization when we have to.