kernel: big virtual memory system overhaul
The primary motivation here was to make the oom killer
more robust and improve consistency in mtaintaining
the invariants.

make killbig() use canqlock(), just in case.

get rid of mcountseg() and maintain a Segment.used and Segment.swapped count,
making it now possible for killbig() to get memory usage data without
needing to acquire the segment lock.

Ensure that when calling putseg(), the original reference from Proc.seg[]
does not point to it. Ensuse that anyone modifying Proc.seg[] must acquire
Proc.seglock. Affected: segdetach(), pexit(), txt2data(), segio.
This will help with debugging.

Make data2txt() take a ref for Image.s when possible, handle release
of original data segment in caller instead.

Fix a A->B->A race with mfreeseg() in fault.c:^pio() function and make sure the
txtflush bits are set *BEFORE* we add the page to the cache.

Dont call procctl() in fault when up->nerrlab > 0, we have to
error() out instead, and run the error handlers!

Improve the per-architecture fault handlers, by having fault()
handle the Porc.insyscall bit restoration and provide a helper
faultnote() that let us avoid allocating a ERRMAX string on the stack,
saving stack space.

Try to free the pages of pte's in batches, avoiding palloc lock contention.
For this, provide deadpage() function that just decrements refcount and
returns non-nil Page when page can be freed. Then caller can chain the
pages and build a freelist to be submitted to freepages() in a batch.