kernel: avoid pagereclaim() on active images when possible
We should not attempt to pagereclaim() active images
while there are still idle ones around.

So do two passes over the image array, and only when
that is exhausted we squeeze active images (which mostly
just yields pages that have been copied-on write
and are now unused from data segment).