kernel: New page cache design
Image.pghash had fixed size, and we
use images with very different sizes.
(fscache).

Most binaries are small, so we are
wasting kernel memory for most images.

So allocate images dynamically for the expected
maximum size: Adding newimage(pages).

Each images gets exactly the pghash size
it needs. Large images (>1024 entries)
get reduced, but with only 16 collisions max.

With that, we can get rid of the collision
check in pagereclaim(), as images should be
balance now.

The next thing is that reclaiming idle
images (binaries) was done randomly.

Instead, we now maintain a orderd "idle list"
of images that is sorted by least-frequent
and decreasing page usage.

We also observe that reclaiming pages from
active images is not worth it. (Unless one
wants to exceute binaries larger than the
available physical memory).

Paging in binaries now uses iounit reads,
and cachepage() ensures there are no duplicate
pages in the cache, not adding a page to the
cache when already present.