git: send more known refs, in hopes of smaller fetches
we used to only say we had the branches we were pulling,
but that meant that when pulling a new branch, we would
say we had nothing, leading to a lot of duplicate data
getting sent.

This change the 20 most recent local branches as refs
we know about, in the hopes that the remote can compute
a smaller delta.

The long term fix would be to write the request in
parallel with reading the response, which would allow
us to provide a great deal more information without
blocking, but this will do the job for now.