kernel: leave alignment hole between TEXT and DATA unmapped
We used to pad the TEXT segment to 2MB aligment on amd64,
resulting in the segment going beyond its image file range.

This broke the new pio(), resulting in an endless loop of
zero length reads of the image.

Instead, we now round segment sizes to our effective
page size only and leave the alignment "hole" unmapped.