kernel: change newpage() prototype
the "clear" argument was almost never used, and
with the new fillpage() function it can be done
as easily by the caller.

Put the virtual address argument "va" first,
it is the most important.

remove the Segment** pointer and just pass a QLock*
as the last argument, signaling we hold locks
and unlock it before sleeping. when that happend,
newpage() now just returns nil to signal we slept
(and lost the lock).