libmemdraw: arbitrary color chans and other improvements for memaffinewarp()
I imported some of the helper routines from draw.c and drawtest.c,
tied them together to make a (get|put)pixel() pair and made an initial
attempt at optimizing them.  Also separated the (bilinear) filtering
code and put it into its own function for modularity (now we can
easily plug any filter we want in the pipeline), along with an example
nearest neighbor one which is disabled for now.

The fixed-point Foint has disappeared in favor of using Point, and
fix_xform() was simplified to avoid unnecessary operations, exploiting
the affine matrix properties and the fact that the transformed point
always has a w coordinate of 1.

Finally, for the inner loop the calls to fix_xform() were replaced by
an incremental method.