libdraw: locking revamp
Guard the display buffer behind an internal QLock
and provide an RWLock instead for the user.  This
makes it possible to draw in parallel.

The behavior by default (when doing the
display->locking=1 and (un)?lockdisplay dance) is
the same, so existing programs won't break.

Update the manuals and document the special
behavior of menuhit(2) and enter(2) when used in
concurrent programs, since they steal the input
entirely and the user has no way of letting them
know when a resize occurs.  Because of this, when
using screen or _screen, or a custom Screen based
on a volatile image, they may become invisible
while kidnapping the input or end up committing a
despicable use-after-free.