libdraw: no more racey Drawops
In draw(2) calls that take a Drawop as an argument
(namely draw [d], line [L], string [sx], ellipse
[eE] and poly [pP]) a message for setting the op
is sent before the actual drawing message, each
using a separate bufimage(2) call.  This caused
issues with programs that draw concurrently, where
a process or thread using an op other than SoverD
would set it, then be rescheduled to another proc
or thread that draws using another op and
overwrite the former's op, which ends up being
forced to use SoverD.  Another case is when the
second bufimage(2) call (the drawing message)
fails for the first proc, then a second one that
doesn't set an op (uses SoverD) will end up
inheriting the outstanding operator.

This changes the procedure to always allocate the
space necessary to hold both the operator setting
and drawing messages when necessary.