c: use hint for register pair
Avoid register allocation leak on 32-bit to 64-bit casts,
by re-using the target register.

This change allows 5c (and hopefully the other ones) to work with more
complex expressions without running out of fixed registers.
The expected byproduct is a minimal reduction in the number of code
size and therefore speed.

A generic stack spill could be better, but this suffices to compile some
programs that make heavy use of 64-bit arithmetic.