upas/fs: handle unsolicited expunge messages correctly
when a different client deletes a message that we haven't added
to the current set of message mid-sync, we would return an error,
leaving the response queued; this would cause the next fetch to
get the response for the previous fetch request, offsetting the
message bodies by one.

Additionally, we were using 'nmsg' and not 'muid' when expunging,
but 'nmsg' could be modified by other unsolicited messages; this
could lead us to indexing out of bounds when moving the messages.