etheriwl: various improvements
- set the interrupt coalescing timer to 2048 usec

- match with 0x08b4: intel 3160

- add a quirk for a hardware bug in 7260

- all quirks meant for 7260 is also meant for 3160

- make sure to set RfhDmaSingleFrame while enabling
RX DMA in mqrx cards. we dont handle multiple frames
per rx buffer yet.

- set the keep alive seconds in setmacpowermode() to
the maximum of 25 and 3*DTIM. a value below 25 results
in weird power saving behavior which results in extremely
slow speeds (25kb/s)

- enable power saving in setmacpowermode(), which allows
the fw to turn off tx/rx circuitry when idle. this should
be fine when associated, and should also be fine while
scanning (as in theory the probe request would wake up
the card and keep it awake for at least a few beacon intervals)

- use Bpool to allocate rx buffers. have a pool of at least 2*Nrx