tcp: splice local tcp conversations together, bypassing ip stack
When we are creating a connection to outselfs, we can skip
tcp and ip completely and just make the conversations into
a pair of pipes.

This makes loopback tcp connections as fast as pipes
and saves memory as no retransmission is needed.

For this, new qsetbypass() function was added, that changes
a (transmission) queue into a bypass at runtime.

It drains whatever is in the queue first and then setting
bypass function pointer. A queue can be un-bypassed again
by passing nil as the bypass function.