devether: eliminate loopback, make it possible to talk to devices sharing our own mac address
Now that loop-back is handled fully by devip itself,
we do not need to consider the "tome" check in transmission
anymore.

This makes it possible to communicate with hosts that share
the same mac address as us (but use differen ip addresses),
which arises with some virtual-machine (ipvtap).

Also remove the same mac check in ethermedium for arp,
we consider this not an error anymore.

Hello sleepydog :)