awk: fix use-after-free in split
when the string passed to split is an element of the array also
passed, e.g.  `split(a[x], a)', we don't want to lose the string value
when reseting the array.  however, the cell is freed during the reset,
and y is now invalid anyway.  instead, detect this case and strdup the
string if needed.  the alternative solution is to always strdup the
string, but this has no practical performance impact even for large
arrays, the intent is clearer and avoids the allocation in all other
cases.