nusb/ether/asix.c: add support for the AX88179A variant
We can tell the 88179A variant from the 88179 by inspecting the USB
device release numbers.  The 88179A by default uses a different
firmware, so we have to set it to use the 88179 compatibility
firmware.

These chips use a slightly different structure for the receive
buffers, where there are "dummy" headers interwoven, each with pktlen == 0.

Also, use ushorts when assigning a value from a GET2().