kernel: fix off-by-one in sysexec progargs
When terminating progargs[], it was possible that
we would set progargs[nelem(progargs)] = nil.

The variable after the progargs array is
volatile char *args, which is already nil,
and not used until later.

Fix the off-by-one, and inline the code for
shargs() directly as it makes it more easy to
see what is going on.

We now raise an error when there are more than
31 (nelem(progargs)-2) arguments passed.