kernel: fix namec() Aunmount semantics for directories
When unmounting a directory from a union, we must not
open it and use the qid from the walk to get the same
qid as was returned by Abind.

This fixes the error one gets when trying to unmound
/mnt/term/dev from /dev, which happens as exportfs
returns a different qid when opend vs walk in its
qid collision scheme.