kernel: refactor devallowed() check in namec() for # devices
Now that we'r not changing amode from Aunmount
to Aopen anymore, we can move the devallowed()
check conditions for Aunmount after parsing # name
and get rid of the "devunmount" conditional.

The utflen(name) == 2 is equivalent to testing
if the attach spec and walk name is empty after
parsing the device character rune.