gefs: wait for epoch before freeing snapdl
freedl() doesn't participate in EBR becasue it assumes
cleaning up a snapshot is only done when a snapshot is
fully inaccessible, but that's not entirely true for
the snap tree -- so we need to wait for all readers
to advance the epoch before we finish the cleanup.

This should be fine, since we're not blocking any of
the mutator procs at this point.